Add files via upload
This contains the basics of numpy arrays, and it's various methods such as how to navigate to an element through it's index, rows & columns, sorting and many more. if you are a beginner in data science, do check out this notebook!
The simultaneous description of functions and methods is what makes this notebook of mine different from others.
make full use of it!